Bổ sung kiểm tra quyền truy cập và cải tiến giao diện xác thực thông tin người sử dụng.
