+++ src/Haihv.Elis.Tool.TraCuuGcn.Api/Extensions/ClaimsPrincipalExtensions.cs
@@ -71,4 +71,5 @@ public static class ClaimsPrincipalExtensions
     {
         return user.GetUserRoles().Contains(role);
     }
+    
 }
