+++ src/Haihv.Elis.Tool.TraCuuGcn.Api/appsettings.json
@@ -47,10 +47,11 @@
   //  "Redis": {
   //    "ConnectionString": "localhost:6379,user=admin,password=admin@Sa-123"
   //  },
-  "JwtTokenSettings": {
-    "SecretKeys": ["SecretKey"],
-    "Issuers": ["https://localhost:5001"],
-    "Audiences": ["https://localhost:5001"]
+
+  "OpenIdConnect": {
+    "Authority": "https://sso.localhost",
+    "Audience": "account",
+    "RoleLocalUser": "domain user"
   },
   "ElisSql": {
     "Databases": [
