+++ src/Haihv.Elis.Tool.TraCuuGcn.Web-App/Haihv.Elis.Tool.TraCuuGcn.Web-App.Client/Services/ClientChuSuDungService.cs
@@ -23,4 +23,17 @@ internal class ClientChuSuDungService(HttpClient httpClient) : IChuSuDungService
             return ([], e.Message);
         }
     }
+
+    public async Task<(bool success, string? error)> GetHasReadPermissionAsync(string serial, string soDinhDanh)
+    {
+        try
+        {
+            var response = await httpClient.GetAsync(ChuSuDungUri.GetHasReadPermissionWithQuery(serial, soDinhDanh));
+            return (response.IsSuccessStatusCode, await response.ParseErrorMessageAsync());
+        }
+        catch (Exception)
+        {
+            return (false, "Lỗi khi kiểm tra quyền truy cập theo số định danh");
+        }
+    }
 }
