+++ src/Haihv.Elis.Tool.TraCuuGcn.WebLib/Components/ThongTinChuSuDung.razor
@@ -59,7 +59,7 @@
 }
 
 @inject IChuSuDungServices ChuSuDungServices
-@inject SnackbarService SnackbarService
+@inject ISnackbar SnackbarService
 
 @code {
     [Parameter] public string? Serial { get; set; }
