+++ src/Haihv.Elis.Tool.TraCuuGcn.WebLib/Components/ThongTinThuaDat.razor
@@ -80,7 +80,7 @@
 }
 
 @inject IThuaDatServices ThuaDatServices
-@inject SnackbarService SnackbarService
+@inject ISnackbar SnackbarService
 
 @code {
     [Parameter] public string? Serial { get; set; }
