+++ src/Haihv.Elis.Tool.TraCuuGcn.WebLib/Components/ThuaDatLocation.razor
@@ -34,7 +34,7 @@
 
 @inject LeafletMapService LeafletMapService
 @inject IThuaDatServices ThuaDatServices
-@inject SnackbarService SnackbarService
+@inject ISnackbar SnackbarService
 
 @code
 {
